+++ packages/amazonq/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.77.0 2025-06-18
+
+- Miscellaneous non-user-facing changes
+
+## 1.76.0 2025-06-18
+
+- **Bug Fix** /transform: only show lines of code statistic in plan
+- **Feature** Add model selection feature
+- **Feature** Pin context items in chat and manage workspace rules
+
 ## 1.75.0 2025-06-13
 
 - Miscellaneous non-user-facing changes